Выполняем ДЗ '5.14. Модуляция', созданы data.js и utils.js. Поправлен main.js, как точка входа.
