+++ js/utils.js
@@ -1,0 +1,36 @@
+//Генератор случайных чисел
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+//Проверка на уникальность id
+
+const getRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      console.error(`Перебраны все числа из диапазона от ${ min } до ${ max }`);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+const generateRandomIdPost = getRandomIdFromRangeGenerator(1, 25);
+const generateRandomIdPhoto = getRandomIdFromRangeGenerator(1, 25);
+const randomCommentId = getRandomIdFromRangeGenerator(1, 5000);
+
+
+export {getRandomInteger};
+export {generateRandomIdPost};
+export {generateRandomIdPhoto};
+export {randomCommentId};
